% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesTLS-package.R
\docType{package}
\name{bayesTLS-package}
\alias{bayesTLS}
\alias{bayesTLS-package}
\title{bayesTLS: Joint Bayesian 4PL Models for Thermal Load Sensitivity}
\description{
Functions for fitting a single joint Bayesian four-parameter logistic
model to thermal-tolerance proportion data and extracting all classical
TLS quantities (z, CTmax at 1 hour, T_crit) and downstream predictions
(heat-injury accumulation, predicted survival under fluctuating
temperature traces, optional Sharpe-Schoolfield repair) from a single
fitted model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daniel1noble/tls_model_equivalence}
  \item Report bugs at \url{https://github.com/daniel1noble/tls_model_equivalence/issues}
}

}
\author{
\strong{Maintainer}: Daniel W. A. Noble \email{daniel.noble@anu.edu.au}

Authors:
\itemize{
  \item Pieter A. Arnold
  \item Patrice Pottier \email{patrice.pottier@bioenv.gu.se}
}

}
\keyword{internal}
