% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clock_to_minutes}
\alias{clock_to_minutes}
\title{Convert various clock formats to minutes}
\usage{
clock_to_minutes(x)
}
\arguments{
\item{x}{Time value(s).}
}
\value{
Numeric vector of minutes since 00:00.
}
\description{
Accepts POSIXt, hms / difftime, numeric fractions of a day (Excel time),
bare numerics (assumed minutes), or \code{"HH:MM:SS"} character strings.
}
\examples{
clock_to_minutes("08:30:00")
clock_to_minutes(0.5)     # half a day = 720 min
}
