% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_4pl.R
\name{has_fit}
\alias{has_fit}
\title{Whether a workflow has been fitted}
\usage{
has_fit(workflow)
}
\arguments{
\item{workflow}{Object returned by \code{\link[=fit_4pl]{fit_4pl()}}.}
}
\value{
Logical scalar.
}
\description{
Whether a workflow has been fitted
}
\examples{
wf <- list(fit = NULL); class(wf) <- "tdt_4pl_workflow"
has_fit(wf)
}
