% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_4pl.R
\name{print.tdt_4pl_workflow}
\alias{print.tdt_4pl_workflow}
\title{Brief printout of a TDT workflow}
\usage{
\method{print}{tdt_4pl_workflow}(x, ...)
}
\arguments{
\item{x}{A \code{tdt_4pl_workflow} object.}

\item{...}{Ignored.}
}
\description{
Brief printout of a TDT workflow
}
