% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snowgum_psii}
\alias{snowgum_psii}
\title{Snow gum leaf PSII functional-impairment thermal-tolerance data}
\format{
A data frame with 319 rows and 8 variables:
\describe{
\item{Temp}{Assay temperature (degrees C).}
\item{Time}{Exposure duration (minutes).}
\item{initial_fvfm}{\eqn{F_v/F_m} measured before heat exposure.}
\item{final_fvfm}{\eqn{F_v/F_m} measured 24 h after heat exposure.}
\item{fvfm_prop}{Retained PSII function, \code{final_fvfm / initial_fvfm}
(a proportion in \link{0, 1}).}
\item{Unique_ID}{Leaf-segment / tree identifier.}
\item{G_Room}{Glasshouse room (grouping factor).}
\item{Day}{Measurement day (grouping factor).}
}
}
\source{
Snow gum leaf PSII TDT assay (Case Study 3). Raw file:
\code{system.file("extdata", "data_function_PSII_TDT_snowgum.csv", package = "bayesTLS")}.
}
\usage{
snowgum_psii
}
\description{
Chlorophyll-fluorescence (\eqn{F_v/F_m}) measurements on excised snow gum
(\emph{Eucalyptus pauciflora} var. \emph{pauciflora}) leaf segments before and
24 h after heat exposure. The response is the continuous proportion
\code{fvfm_prop} (post/pre ratio), modelled with a Beta likelihood. The
model-ready frame for Case Study 3 (sublethal, continuous-proportion endpoint).
}
\examples{
std <- standardize_data(snowgum_psii, temp = "Temp", duration = "Time",
                        proportion = "fvfm_prop",
                        random_effects = c("Day", "G_Room"),
                        duration_unit = "minutes")
}
\keyword{datasets}
